make the animations reusable
